+++ examples/usb_demo/makefile
@@ -18,7 +18,7 @@ F_CPU        = 16000000
 F_USB        = $(F_CPU)
 OPTIMIZATION = s
 TARGET       = mooltipass
-SRC          = $(TARGET).c usb.c Descriptors.c $(LUFA_SRC_USB)
+SRC          = $(TARGET).c usb.c Descriptors.c usb_cmd_parser.c $(LUFA_SRC_USB)
 #LUFA_PATH    = ../../../LUFA/LUFA/LUFA/
 LUFA_PATH = ../../../LUFA140302/LUFA/
 CC_FLAGS     = -DINTERRUPT_CONTROL_ENDPOINT -DUSE_LUFA_CONFIG_HEADER -IConfig/
